refactor folders structure for streaming tests
