+++ app/services/platforms/facebook.ts
@@ -198,6 +198,7 @@ export class FacebookService extends BasePlatformService<IFacebookServiceState>
 
     // send selected pageId to streamlabs.com
     if (fbOptions.destinationType === 'page') {
+      assertIsDefined(fbOptions.pageId);
       await this.postPage(fbOptions.pageId);
     }
   }
