+++ test/regular/streaming/streaming.ts
@@ -35,174 +35,6 @@ import { StreamSettingsService } from '../../../app/services/settings/streaming'
 
 useSpectron();
 
-test('Streaming to Twitch without auth', async t => {
-  const userInfo = await reserveUserFromPool(t, 'twitch');
-
-  await showSettings(t, 'Stream');
-
-  // This is the twitch.tv/slobstest stream key
-  await setFormInput(t, 'Stream key', userInfo.streamKey);
-  await t.context.app.client.click('button=Done');
-
-  // go live
-  await prepareToGoLive(t);
-  await clickGoLive(t);
-  await waitForStreamStart(t);
-
-  t.pass();
-});
-
-test('Streaming to Twitch', async t => {
-  await logIn(t, 'twitch');
-  await goLive(t, {
-    title: 'SLOBS Test Stream',
-    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
-  });
-  t.true(await chatIsVisible(t), 'Chat should be visible');
-
-  // check we can't change stream setting while live
-  await showSettings(t, 'Stream');
-  await t.true(
-    await t.context.app.client.isExisting("div=You can not change these settings when you're live"),
-    'Stream settings should be not visible',
-  );
-  t.pass();
-});
-
-test('Streaming to a Facebook Page', async t => {
-  await logIn(t, 'facebook');
-  await goLive(t, {
-    title: 'SLOBS Test Stream',
-    facebookGame: selectTitle('Fortnite'),
-    description: 'SLOBS Test Stream Description',
-  });
-  t.true(await chatIsVisible(t), 'Chat should be visible');
-  t.pass();
-});
-
-// TODO: setup FB permissions in the user-pool to run this test
-test.skip('Streaming to a Facebook User`s timeline', async t => {
-  await logIn(t, 'facebook');
-  await goLive(t, {
-    title: 'SLOBS Test Stream',
-    facebookGame: selectTitle('Fortnite'),
-    description: 'SLOBS Test Stream Description',
-    destinationType: 'me',
-  });
-  t.pass();
-});
-
-// TODO: setup FB permissions in the user-pool to run this test
-test.skip('Streaming to a Facebook User`s group', async t => {
-  await logIn(t, 'facebook');
-  await goLive(t, {
-    title: 'SLOBS Test Stream',
-    facebookGame: selectTitle('Fortnite'),
-    description: 'SLOBS Test Stream Description',
-    destinationType: 'group',
-  });
-  t.pass();
-});
-
-test('User does not have Facebook pages', async t => {
-  await logIn(t, 'facebook', { noFacebookPages: true, notStreamable: true });
-  await prepareToGoLive(t);
-  await clickGoLive(t);
-  if (await t.context.app.client.isExisting('button=Go Live')) {
-    await t.context.app.client.click('button=Go Live');
-  }
-  await focusChild(t);
-  t.true(
-    await t.context.app.client.isExisting('a=create a Gaming Video Creator Page'),
-    'The link for adding new facebook changes should exist',
-  );
-});
-
-test('Streaming to Youtube', async t => {
-  await logIn(t, 'youtube');
-
-  t.false(await chatIsVisible(t), 'Chat is not visible for YT before stream starts');
-
-  await goLive(t, {
-    title: 'SLOBS Test Stream',
-    description: 'SLOBS Test Stream Description',
-  });
-
-  t.true(await chatIsVisible(t), 'Chat should be visible');
-
-  // give youtube 2 min to publish stream
-  await focusChild(t);
-  await t.context.app.client.waitForVisible("h1=You're live!", 2 * 60 * 1000);
-
-  t.pass();
-});
-
-test('Streaming to the scheduled event on Youtube', async t => {
-  await logIn(t, 'youtube', { multistream: false });
-
-  // create event via scheduling form
-  const tomorrow = Date.now() + 1000 * 60 * 60 * 24;
-  const formattedTomorrow = moment(tomorrow).format(moment.localeData().longDateFormat('ll'));
-  await scheduleStream(t, tomorrow, {
-    title: 'Youtube Test Stream',
-    description: 'SLOBS Test Stream Description',
-  });
-
-  // select event and go live
-  await prepareToGoLive(t);
-  await clickGoLive(t);
-  const form = new FormMonkey(t);
-  await form.fill({
-    event: await form.getOptionByTitle('event', `Youtube Test Stream (${formattedTomorrow})`),
-  });
-  await submit(t);
-  await waitForStreamStart(t);
-  t.pass();
-});
-
-// TODO: delete all the scheduled on the user-pool and enable this test
-test.skip('Streaming to the scheduled event on Facebook', async t => {
-  await logIn(t, 'facebook', { multistream: false });
-
-  // create event via scheduling form
-  const tomorrow = Date.now() + 1000 * 60 * 60 * 24;
-  const title = `facebook scheduled stream ${tomorrow}`;
-  await scheduleStream(t, tomorrow, {
-    title,
-    description: 'Facebook Test Stream Description',
-  });
-
-  // select event and go live
-  await prepareToGoLive(t);
-  await clickGoLive(t);
-  const form = new FormMonkey(t);
-  await form.fill({
-    fbEvent: selectTitle(title),
-  });
-  await submit(t);
-  await waitForStreamStart(t);
-  t.pass();
-});
-
-test('Start stream twice to the same YT event', async t => {
-  await logIn(t, 'youtube', { multistream: false });
-
-  // create event via scheduling form
-  const now = Date.now();
-  await goLive(t, {
-    title: `Youtube Test Stream ${now}`,
-    description: 'SLOBS Test Stream Description',
-    enableAutoStop: false,
-  });
-  await stopStream(t);
-
-  await goLive(t, {
-    event: selectTitle(`Youtube Test Stream ${now}`),
-    enableAutoStop: true,
-  });
-  t.pass();
-});
-
 test('Stream after switching accounts', async t => {
   // stream to youtube
   await logIn(t, 'youtube');
@@ -251,96 +83,6 @@ test('Stream with disabled confirmation', async t => {
   t.pass();
 });
 
-test('Migrate the twitch account to the protected mode', async t => {
-  await logIn(t, 'twitch');
-
-  // change stream key before go live
-  const streamSettings = (await getClient()).getResource<StreamSettingsService>(
-    'StreamSettingsService',
-  );
-  streamSettings.setSettings({ key: 'fake key', protectedModeMigrationRequired: true });
-
-  await restartApp(t); // restarting the app should call migration again
-
-  // go live
-  await tryToGoLive(t, {
-    title: 'SLOBS Test Stream',
-    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
-  });
-  await waitForStreamStop(t); // can't go live with a fake key
-
-  // check that settings have been switched to the Custom Ingest mode
-  await showSettings(t, 'Stream');
-  t.true(
-    await t.context.app.client.isVisible('button=Use recommended settings'),
-    'Protected mode should be disabled',
-  );
-
-  // use recommended settings
-  await t.context.app.client.click('button=Use recommended settings');
-  // setup custom server
-  streamSettings.setSettings({
-    server: 'rtmp://live-sjc.twitch.tv/app',
-    protectedModeMigrationRequired: true,
-  });
-
-  await restartApp(t); // restarting the app should call migration again
-  await tryToGoLive(t, {
-    title: 'SLOBS Test Stream',
-    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
-  });
-  await waitForStreamStop(t);
-
-  // check that settings have been switched to the Custom Ingest mode
-  await showSettings(t, 'Stream');
-  t.true(
-    await t.context.app.client.isVisible('button=Use recommended settings'),
-    'Protected mode should be disabled',
-  );
-});
-
-test('Schedule stream to facebook', async t => {
-  // login into the account:
-  await logIn(t, 'facebook', { multistream: false });
-  const app = t.context.app;
-
-  // open EditStreamInfo window
-  await focusMain(t);
-  await app.client.click('button .icon-date');
-  await focusChild(t);
-  const formMonkey = new FormMonkey(t);
-
-  // wait fields to be shown
-  await app.client.waitForVisible('[data-name=title]');
-
-  // fill streaming data
-  await formMonkey.fill({
-    title: 'SLOBS Test Stream',
-    description: 'SLOBS Test Stream Description',
-  });
-
-  // set the date to tomorrow
-  const today = new Date();
-  const tomorrow = new Date();
-  tomorrow.setDate(today.getDate() + 1);
-  await formMonkey.fill({
-    date: moment(tomorrow).format('MM/DD/YYYY'),
-  });
-  await app.client.click('button=Done');
-
-  // facebook requires a game
-  await app.client.waitForVisible('.toasted.error', 2000);
-
-  await sleep(3000); // there are some issues with setting the game field without delay here
-  await formMonkey.fill({
-    game: selectTitle('Fortnite'),
-  });
-
-  await app.client.click('button=Done');
-  await app.client.waitForVisible('.toast-success', 30000);
-  t.pass();
-});
-
 test('Go live error', async t => {
   // login into the account
   await logIn(t, 'twitch');
@@ -373,15 +115,6 @@ test('Go live error', async t => {
   await app.client.waitForVisible('a=just go live');
 
   t.pass();
-});
-
-test('Youtube streaming is disabled', async t => {
-  skipCheckingErrorsInLog();
-  await logIn(t, 'youtube', { streamingIsDisabled: true, notStreamable: true });
-  t.true(
-    await t.context.app.client.isExisting('span=YouTube account not enabled for live streaming'),
-    'The streaming-disabled message should be visible',
-  );
 });
 
 test('User has linked twitter', async t => {
@@ -447,63 +180,5 @@ test('Streaming to Dlive', async t => {
 test('Update channel settings before streaming', async t => {
   await logIn(t, 'twitch');
   await updateChannelSettings(t, { title: 'updated title' });
-  t.pass();
-});
-
-test('Custom stream destinations', async t => {
-  const client = t.context.app.client;
-  await logIn(t, 'twitch', { prime: true });
-
-  // fetch a new stream key
-  const user = await reserveUserFromPool(t, 'twitch');
-
-  // add new destination
-  await showSettings(t, 'Stream');
-  await click(t, 'span=Add Destination');
-  await fillForm(t, null, {
-    name: 'MyCustomDest',
-    url: 'rtmp://live.twitch.tv/app/',
-    streamKey: user.streamKey,
-  });
-  await click(t, 'button=Save');
-  await t.true(await client.isExisting('span=MyCustomDest'), 'New destination is created');
-
-  // update destinations
-  await click(t, 'i.fa-pen');
-  await fillForm(t, null, {
-    name: 'MyCustomDestUpdated',
-  });
-  await click(t, 'button=Save');
-  await t.true(await client.isExisting('span=MyCustomDestUpdated'), 'Destination is updated');
-
-  // add one more destination
-  await click(t, 'span=Add Destination');
-  await fillForm(t, null, {
-    name: 'MyCustomDest',
-    url: 'rtmp://live.twitch.tv/app/',
-    streamKey: user.streamKey,
-  });
-  await click(t, 'button=Save');
-  await t.false(
-    await client.isExisting('span=Add Destination'),
-    'Do not allow more than 2 custom dest',
-  );
-
-  // open the GoLiveWindow and check destinations
-  await prepareToGoLive(t);
-  await clickGoLive(t);
-  await t.true(await client.isExisting('span=MyCustomDest'), 'Destination is available');
-  await click(t, 'span=MyCustomDest'); // switch the destination on
-  await tryToGoLive(t);
-  await client.waitForExist('span=Configure the Multistream service'); // the multistream should be started
-  await stopStream(t);
-  await releaseUserInPool(user);
-
-  // delete existing destinations
-  await showSettings(t, 'Stream');
-  await click(t, 'i.fa-trash');
-  await click(t, 'i.fa-trash');
-  t.false(await client.isExisting('i.fa-trash'), 'Destinations should be removed');
-
   t.pass();
 });
