+++ test/regular/streaming/twitch.ts
@@ -1,0 +1,100 @@
+import { restartApp, test, useSpectron } from '../../helpers/spectron';
+import { logIn, reserveUserFromPool } from '../../helpers/spectron/user';
+import { showSettings } from '../../helpers/spectron/settings';
+import { setFormInput } from '../../helpers/spectron/forms';
+import {
+  chatIsVisible,
+  clickGoLive,
+  goLive,
+  prepareToGoLive,
+  tryToGoLive,
+  waitForStreamStart,
+  waitForStreamStop,
+} from '../../helpers/spectron/streaming';
+import { selectTitle } from '../../helpers/form-monkey';
+import { getClient } from '../../helpers/api-client';
+import { StreamSettingsService } from '../../../app/services/settings/streaming';
+
+useSpectron();
+
+test('Streaming to Twitch without auth', async t => {
+  const userInfo = await reserveUserFromPool(t, 'twitch');
+
+  await showSettings(t, 'Stream');
+
+  // This is the twitch.tv/slobstest stream key
+  await setFormInput(t, 'Stream key', userInfo.streamKey);
+  await t.context.app.client.click('button=Done');
+
+  // go live
+  await prepareToGoLive(t);
+  await clickGoLive(t);
+  await waitForStreamStart(t);
+
+  t.pass();
+});
+
+test('Streaming to Twitch', async t => {
+  await logIn(t, 'twitch');
+  await goLive(t, {
+    title: 'SLOBS Test Stream',
+    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
+  });
+  t.true(await chatIsVisible(t), 'Chat should be visible');
+
+  // check we can't change stream setting while live
+  await showSettings(t, 'Stream');
+  await t.true(
+    await t.context.app.client.isExisting("div=You can not change these settings when you're live"),
+    'Stream settings should be not visible',
+  );
+  t.pass();
+});
+
+test('Migrate the twitch account to the protected mode', async t => {
+  await logIn(t, 'twitch');
+
+  // change stream key before go live
+  const streamSettings = (await getClient()).getResource<StreamSettingsService>(
+    'StreamSettingsService',
+  );
+  streamSettings.setSettings({ key: 'fake key', protectedModeMigrationRequired: true });
+
+  await restartApp(t); // restarting the app should call migration again
+
+  // go live
+  await tryToGoLive(t, {
+    title: 'SLOBS Test Stream',
+    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
+  });
+  await waitForStreamStop(t); // can't go live with a fake key
+
+  // check that settings have been switched to the Custom Ingest mode
+  await showSettings(t, 'Stream');
+  t.true(
+    await t.context.app.client.isVisible('button=Use recommended settings'),
+    'Protected mode should be disabled',
+  );
+
+  // use recommended settings
+  await t.context.app.client.click('button=Use recommended settings');
+  // setup custom server
+  streamSettings.setSettings({
+    server: 'rtmp://live-sjc.twitch.tv/app',
+    protectedModeMigrationRequired: true,
+  });
+
+  await restartApp(t); // restarting the app should call migration again
+  await tryToGoLive(t, {
+    title: 'SLOBS Test Stream',
+    twitchGame: selectTitle("PLAYERUNKNOWN'S BATTLEGROUNDS"),
+  });
+  await waitForStreamStop(t);
+
+  // check that settings have been switched to the Custom Ingest mode
+  await showSettings(t, 'Stream');
+  t.true(
+    await t.context.app.client.isVisible('button=Use recommended settings'),
+    'Protected mode should be disabled',
+  );
+});
